Change enqueuing items to use Add instead of AddRateLimited.
Calling 'AddRateLimited' appears to indicate that an error has occurred
for the rate limiting queue.  This triggers exponential backoff
behavior, so if a large number of elements (20+) are owned by a resource
and the controller happens to encounter an error, this will cause
exponential backoff to reach the 1000 second limit.