Plugin model; --pattern=addon flag that generates an addon operator
We implement the initial plugin model, starting with an in-process
implementation that can support out-of-process later.  As plugins do
not yet form part of the stable interface for kubebuilder, we only
enable them when the KUBEBUILDER_ENABLE_PLUGINS env var is set.

We implement an initial plugin for addons: `--pattern=addon` creates a
controller & resource that follow the style of
https://github.com/kubernetes-sigs/kubebuilder-declarative-pattern

This style of declarative addon operators is being investigated in the
cluster-addons subject, with code at
https://github.com/kubernetes-sigs/addon-operators