Use gofmt, not goimports
goimports is slower, should not be necessary (we don't want to add
other imports), and the order of import declarations is no longer
consistent, because goimport groups them alphabetically.

goimports also means that the behaviour of kubebuilder depends on what
is on the GOPATH.

However, goimports will also respect existing import blocks, so we
remain compatible with goimports.