A CLI is now able to resolve a key to one plugin for a project version containing multiple plugins of the same name (different versions) or same short name (different fully-qualified names).
For example, when a CLI containing plugins "go.kubebuilder.io/v2.0"
and "go.kubebuilder.io/v2.1" with a default of the latter is passed
a either config with "layout: go/v2.0", a config with "layout: go/v2.1",
or "--plugins go", it should resolve each successfully.

pkg/cli: resolve default plugins first so that, if a key matches a
default plugin, the correct plugin is always returned. If a user wants
non-default behavior they must specify the exact plugin they want.