Fix book on dark theme devices & update mdbook
We were using a index template from a version of mdbook with support for
autodetecting devices/browser's dark theme preferences, but we didn't
have the corresponding mdbook binary, which was causing the site to
break on devices that prefer dark theme.

This updates mdbook, as well as updating the index template to
more-or-less match the latest changes with our custom patches (custom
favicon path, custom header location, custom header image).