:bug: Don't recurse into hidden directories
checkDir() currently recurses into hidden directories at the root. This
is a problem, because only the directory itself will have a . at the
root of it's path, but any files within it won't when calling info.Name.
This causes init to fail on a directory containing any hidden directory
with files within them, such as .git.

Since we use filepath.Walk(".") we can't skip a directory that just
starts with ., as that would include the current directory. So instead
we also check that Name() doesn't return just "."