Fix advice on using manager's client in tests
Previous documentation erroneously misconfigured the cronjob's tutorial
to use the manager's client everywhere, which is poor practice because
using a cache client in assertions leads to flaky & slow tests.

A new PR overcorrected, and switched both assertions *and* the
reconciler to use a live client, which is also wrong, because the
reconciler needs to use a manager's client to function properly.

This corrects the documentation to indicate that reconcilers should use
managers' clients, and test assertions should use a live client, and
adds a full explanation as to why.