:bug: add the path to Kubebuilder assets in go/v3 (#2272)
Currently, while setting up envtest and running controller
tests in go/v3 projects, we look for the binary in the default
location. With the changes made in `setup-envtest`, the kb assets
are downloaded at tmp/k8s/<os_specific_location>. This fix sets
the KUBEBUILDER_ASSETS var to point to the location where they
are downloaded so that `make test` does not fail.

Signed-off-by: varshaprasad96 <varshaprasad96@gmail.com>