stop to generate crd webhooks patches and cainjetions for any CRD/API and projects without webhooks
At present, we scaffold config/crd/patches, kustomizations, and CA injections for every CRD, irrespective of whether webhooks are enabled in the project or not. However, these configurations are only relevant and valid if the project has webhooks.

Consequently, for projects without webhooks, this leads to failures as documented in kubebuilder pull request #3585.

To address this, we are now introducing a test to ensure that projects without enabled webhooks function correctly and as anticipated.

Signed-off-by: Camila Macedo <7708031+camilamacedo86@users.noreply.github.com>
Co-authored-by: lowang-bh <lhui_wang@163.com>