Clean up getting started docs
Add generation of getting started project code

This adds a step to generate the code used in the Getting Started docs
page. This allows us to reference the pre-generated code relative to the
docs page so we can keep things in sync between the generated code and
the examples given in the docs.

Fix various typos and grammar in v1alpha1 template

Addresses various spelling and grammar mistakes found in the
template used to generate the v1alpha1 sample projects.

Addresses various typos, grammatical errors, and other refinement
in the Getting Started docs page.

Where possible, source files from the generated testdata project are
imported directly in the docs so they stay up to date. In other areas
where it wasn't practical to pull in the entire source file, this
updates the code snippet shown to match the generated source and
provides a link to the GitHub repo to use as a reference.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>