🐛 fix place where metrics service is scaffolded by moving from config/rbac to config/default (#3945)
fix place where metrics service is scaffolded by moving from config/rbac to config/default

When we discontinued the usage of kube-rbac-proxy we placed the Metrics Service under config/rbac but it is not the best place to fit this resource. Furthermore, within those changes we are ensuring that the metrics service will only be applied if/when users enable the metrics.