📖 Upgrade controller-gen used to generated the docs and remove redirects for artefact images
This PR update the deps used to generate the data in the docs.
It should update the markers to address https://github.com/kubernetes-sigs/kubebuilder/issues/4009

Furthermore, for we are able to move forward here we need to remove the redirect for `https://go.kubebuilder.io/images` and `gcr.io/kubebuilder/thirdparty` which are deprecated and/or no longer used.

We are either adding a note in the in the Artefacts page to clarify that ENV TEST binaries are no longer produced by Kubebuilder and its build and promotion was moved to controller-gen.

Close: https://github.com/kubernetes-sigs/kubebuilder/issues/4009