:book: update the next step from quick start accordingly
Now that we have the Getting Started guide, which provides a solid overview, the most appropriate next step seems to be following this guide. Lastly, we are removing the note from the DeployImage section as it doesn't seem to fit well on this page.

See that now we have the Getting Started which provides a good overview
so the most approprieted next step shows to be this one. By last, we are removing the note from DeployImage it does not seems to fit well in this page