🌱 (ci) - Fix the job to check the PR title and no longer use deprecated github action (#4095)
The kubebuilder-release-tools project is no longer really maintained. The image required for checks is deprecated due
the infrastructure used to build and promote it be deprecated. Given the minimal nature of the check, it's unnecessary to maintain a GitHub Action for this purpose.

Also, it broke 1 day ago. Note that we need to pass permissions via GitHub Token, which is not very safe. This is another reason why we need to proceed with these changes.