:bug: fix: Enhance the webhook test template
The webhook scaffold had some oddly indented regions, resulting in the
generated source code having a mix of tabs and spaces, and was generally
looking quite ugly.

The gomega asserions are too chatty; errors, and return values are
checked after first being assigned as variables in scope. This is not
needed, and Gomega checks all of this for us:

Expect(someFunc()).Error().To(HaveOccurred()) checks both the error, but also that the other return values are zero or nil.
Expect(someFunc()).To(ContainSubstring("X")) checks both the return value and verifies that the error is nil.

- Indentation
- simplify gomega assertions