cleanup: simplify and consolidate testdata samples #4126
The testdata samples are designed to showcase various options for users and enable comprehensive testing of tool commands. By aggregating the Grafana and deploy-image plugins into the multigroup layout, we streamline test coverage and reduce redundancy. This change ensures all options are tested within the multigroup layout while keeping project-v4 with and without the optional plugins, to help us ensure the references for users and keep our CI tests.

Now, instead of having a separate sample for each optional plugin, we have:

- project-v4
- project-v4-multigroup-with-plugins
- project-v4-with-plugins

This consolidation reduces the number of samples and tests, making the codebase easier to maintain without sacrificing test coverage or usability for users.