fix: improve CronJob and multiversion tutorial by refining replaces without overwriting webhook markers
Ensure we are more precise with  operations to avoid overwriting the default scaffolding. Previously, the markers in the scaffold were unintentionally overwritten when injecting code and documentation, which could override crucial changes.

This commit ensures the fix applied in PR #4122 is preserved by refining the tutorial instructions to avoid overwriting webhook markers while still injecting the necessary information.

Either we should not export the consts used in the hack/docs since those are internal implementations, therefore the consts used in this area affected have their name changed