:bug: fix: replace []byte with string in utils.Run()
utils.go has a function to execute a command, and it returns the output
if the command is successful. It returns a byte array, the same as the
underlying exec.Cmd() provides. This, being a test heper function would
make sense to return a string instead, so that the ergonomics of using
Gomega assertions on strings work better by default.

Removed the now unneccesary cast from []byte to string.

Found an additional place where byte arrays were being asserted against
ContainsSubstring (plugin_cluser_test.go).