clarify usage of alpha generate command as internal-only
Added comments to `NewScaffoldCommand` to clarify that the `kubebuilder alpha generate` command is designed for Kubebuilder’s internal use only.

The updated go documentation explains that importing this command into other projects is not supported, as it relies on Kubebuilder-specific configurations, key mappings, and plugins, which may not be applicable to other setups.

Projects that use Kubebuilder as a library and require similar functionality should implement a custom solution for compatibility with their specific configurations and plugins.