revert the changes done in the PR #4286
We are reverting the changes done in the PR #4286.
We are just keeping the e2e tests improvements.
We will not able to solve the issue https://github.com/kubernetes-sigs/kubebuilder/issues/4146 without add new markers.

We cannot skip the action to write an webhook since
it will result in incomplete inplementation in the <kind>_webhook.go as in
the <kind>_webhook_test.go